+++ src/backend/mi_parse.ts
@@ -78,7 +78,7 @@ export class MINode implements MIInfo {
 			if (target) {
 				path = path.substr(target[0].length);
 				if (current.length && typeof current !== 'string') {
-					const found :string[] = [];
+					const found: string[] = [];
 					for (const element of current) {
 						if (element[0] === target[1]) {
 							found.push(element[1]);